requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-Cors==4.0.0

# AI API
anthropic==0.18.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Network Requests
requests==2.31.0

# Visualization
pyvis==0.3.2
networkx==3.2.1

# Environment Variables Management
python-dotenv==1.0.0

# Tools
python-slugify==8.0.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
